[expo-updates][android][6/n] Create IUpdatesController and split controllers depending on configuration
