+++ packages/expo-updates/android/src/main/java/expo/modules/updates/manifest/ManifestMetadata.kt
@@ -26,7 +26,7 @@ object ManifestMetadata {
   ): JSONObject? {
     return try {
       val jsonString = database.jsonDataDao()!!
-        .loadJSONStringForKey(key, configuration.scopeKey!!)
+        .loadJSONStringForKey(key, configuration.scopeKey)
       if (jsonString != null) JSONObject(jsonString) else null
     } catch (e: Exception) {
       Log.e(TAG, "Error retrieving $key from database", e)
@@ -62,7 +62,7 @@ object ManifestMetadata {
     value: String?
   ) {
     // this is done within a transaction to ensure consistency
-    database.jsonDataDao()!!.updateJSONStringForKey(EXTRA_PARAMS_KEY, configuration.scopeKey!!) { previousValue ->
+    database.jsonDataDao()!!.updateJSONStringForKey(EXTRA_PARAMS_KEY, configuration.scopeKey) { previousValue ->
       val jsonObject = previousValue?.let { JSONObject(it) }
       val extraParamsToWrite = (jsonObject?.asStringStringMap()?.toMutableMap() ?: mutableMapOf()).also {
         if (value != null) {
@@ -93,7 +93,7 @@ object ManifestMetadata {
       fieldsToSet[MANIFEST_FILTERS_KEY] = responseHeaderData.manifestFilters.toString()
     }
     if (fieldsToSet.isNotEmpty()) {
-      database.jsonDataDao()!!.setMultipleFields(fieldsToSet, configuration.scopeKey!!)
+      database.jsonDataDao()!!.setMultipleFields(fieldsToSet, configuration.scopeKey)
     }
   }
